Java Input And Output Techniques
In this program I have shown how you can print the data in java in different ways:
--Using println to print data in next line and prints same as printf("\n")
--Using format specifers like in c language i.e. printf("%d")
--Using simple print to print in same line


How System.out.println works??
System.out.println is a very common command in Java that’s used to print information to the console. When you see this command, it’s helpful to break it down into its components to understand what each part does:

System: This is a predefined final class that provides access to the system, and is part of the java.lang package.
out: This is a static member field of the System class and is an instance of the PrintStream class. It represents the standard output stream (usually the console).
println: This is a method of the PrintStream class. When you call println, it sends the argument you’ve passed to it to the standard output stream, followed by a newline character.
So, when you write System.out.println("Hello, world!");, you’re telling Java to print “Hello, world!” to the console and then move to the next line.
 The System class is part of the core Java library, so you don’t need to create an instance of it to use out or println; they can be accessed directly because out is a static field, and println is a method that can be invoked on this static field.