Java Constructor and objects
Class Definition:
The code defines a class named Student.
This class encapsulates information about a student, including their name, roll number (rn), and percentage (perc).
The class has three private instance variables: name, rn, and perc.

Constructor:
A constructor is a special method that is invoked when an object of the class is created.
Constructors have the same name as the class and do not have a return type.

In this code, the Student class has a parameterized constructor The constructor accepts three parameters: name, rn, and perc.
Inside the constructor, the instance variables are initialized with the provided values using the this keyword.
For example, when you create a Student object, you can pass values like "Kartik", 17090, and 90.5 to initialize the student’s details.

Methods:
The Student class has two methods:
Setdetails(String name, int rn, double perc): This method allows you to update the student’s details after the object is created. It sets the instance variables to the provided values.

Main Class (Mystudent):
The Mystudent class contains the main method, which serves as the entry point for the program.
Inside main, an instance of the Student class is created
The constructor is called with the provided arguments, initializing the student’s details.
Finally, the Displaydata() method is invoked on s1, printing the student’s information to the console.